Minor changes to load_context()

+++ trunk/malloc/malloc.h
@@ -15,7 +15,7 @@
 struct memory_control_block
 {
 	int is_available;
-	UINT16 size;
+	int size;
 };
 
 /* Initialize malloc */
